requirements.txt
discord.py==2.6.3
python-dotenv==1.0.1
aiosqlite==0.20.0
